+++ terraform/modules/helm/helm.tf
@@ -29,12 +29,6 @@ locals {
     "oci-onm-mgmt-agent.mgmtagent.installKeyFileContent" = var.mgmt_agent_install_key_content
     "oci-onm-mgmt-agent.deployMetricServer"              = var.opt_deploy_metric_server
   }
-
-  mushop_helm_inputs = {
-    # oci-onm-logan
-    "createServiceAccount" = false
-    "serviceAccount"       = var.livelab_service_account
-  }
 }
 
 # Create helm release
@@ -59,7 +53,7 @@ resource "helm_release" "oci-kubernetes-monitoring" {
   }
 
   dynamic "set" {
-    for_each = var.deploy_mushop_config ? local.mushop_helm_inputs : {}
+    for_each = var.oci_domain == null ? {} : { "oci-onm-logan.ociDomain" = var.oci_domain }
     content {
       name  = set.key
       value = set.value
@@ -90,7 +84,7 @@ data "helm_template" "oci-kubernetes-monitoring" {
   }
 
   dynamic "set" {
-    for_each = var.deploy_mushop_config ? local.mushop_helm_inputs : {}
+    for_each = var.oci_domain == null ? {} : { "oci-onm-logan.ociDomain" = var.oci_domain }
     content {
       name  = set.key
       value = set.value