+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2024-11-20
+### Added
+- Support for new OCI Regions which are not yet supported through OCI Ruby SDK by default.
+- Status check for OKE lifecycle state to be active before installing helm chart, when installed using RMS.
+- An option to apply a static delay before installing helm chart, when installed using RMS.
+
 ## 2024-11-05
 ### Added
 - Support of extraEnv for Management Agent